tests(firebase_analytics): switch driver tests to use e2e
