+++ packages/firebase_analytics/firebase_analytics/example/test_driver/firebase_analytics_e2e_test.dart
@@ -1,0 +1,7 @@
+// Copyright 2020, the Chromium project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+import 'package:e2e/e2e_driver.dart' as e2e;
+
+void main() => e2e.main();
